python/pyproject.toml
[tool.poetry]
name = "mushroom-field-guide"
version = "0.1.0"
description = "A mushroom field guide data processor using Mushroom Observer data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
motor = "3.3.1"
pymongo = "4.5.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.3"
PyYAML = "^6.0.1"
typing-extensions = "^4.8.0"
aiohttp = "^3.9.1"
aiofiles = "^23.2.1"
tqdm = "^4.66.1"
jsonschema = "^4.23.0"
certifi = "^2025.1.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["motor.*", "pandas.*", "yaml.*"]
ignore_missing_imports = true
